+++ src/__tests__/DetailedCard.test.tsx
@@ -1,4 +1,4 @@
-import { DetailedCard } from '../components/detailedCard/DetailedCard';
+import { DetailedCard } from '../components';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
