+++ src/__tests__/Theme.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { THEMES } from '../context/constants';
-import { Theme } from '../context/Theme';
-import { ThemeContext } from '../context/ThemeContext';
-import s from '../context/style.module.css';
+import { THEMES } from '../app/context/constants';
+import { Theme } from '../app/context/Theme';
+import { ThemeContext } from '../app/context/ThemeContext';
+import s from '../app/context/style.module.css';
 
 const localStorageMock = (() => {
   let store: Record<string, string> = {};
