+++ src/components/card/style.module.css
@@ -1,4 +1,4 @@
 .checkbox {
-    width: 20px;
-    height: 20px;
+  width: 20px;
+  height: 20px;
 }
