+++ src/components/cardList/style.module.css
@@ -52,7 +52,7 @@
 }
 
 .darkTheme.table tbody tr:nth-child(even) {
-  background-color: #9DB2BF;
+  background-color: #9db2bf;
 }
 
 .darkTheme.table tbody tr:hover {
