+++ src/components/detailedCard/style.module.css
@@ -19,5 +19,5 @@
 }
 
 .darkTheme.details {
-    background-color: #b4c2cc;
+  background-color: #b4c2cc;
 }
