+++ src/components/errorBoundary/style.module.css
@@ -1,0 +1,4 @@
+.container {
+  padding: 20px;
+  text-align: center;
+}
