+++ src/components/index.ts
@@ -1,0 +1,14 @@
+export { Button } from './button/Button';
+export { Card } from './card/Card';
+export { CardList } from './cardList/CardList';
+export { DetailedCard } from './detailedCard/DetailedCard';
+export { ErrorBoundary } from './errorBoundary/ErrorBoundary';
+export { Header } from './header/Header';
+export { Input } from './input/Input';
+export { Layout } from './layout/Layout';
+export { Pagination } from './pagination/Pagination';
+export { Page404 } from './page404/Page404';
+export { Search } from './search/Search';
+export { SelectedItems } from './selectedItems/SelectedItems';
+export { Spinner } from './spinner/Spinner';
+export { ThemeToggle } from './themeToggle/ThemeToggle';
