--- src/context/style.module.css
@@ -1,16 +1,0 @@
-.lightThemeContainer {
-  --background: #DBD3C9;
-  --primary-color: #88A0A8;
-  --second-color: #546A76;
-}
-
-.darkThemeContainer {
-  --background: #4E4D5C;
-  --primary-color: #243B4A;
-  --second-color: #805E73;
-}
-
-.themeContainer {
-  min-height: 100vh;
-  background-color: var(--background);
-}
